+++ vfeed/__init__.py
@@ -3,4 +3,5 @@ from .info import vFeedInfo
 from .exportxml import vFeedXML
 from .update import vFeedUpdate
 from .stats import vFeedStats
+from .search import vFeedSearch
 from . import config